refactor: reorganize AuthUpgradeDetail type and improve ProjectOrgSwitcher styling
